requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
plotly==5.14.0
faiss-cpu==1.7.3
umap-learn==0.5.3 